requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==1.3.0
django==4.2.24
django-appconf==1.0.6
django-imagekit==5.0.0
gunicorn==23.0.0
packaging==25.0
pilkit==3.0
pillow==10.4.0
psycopg2-binary==2.9.10
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
reportlab==3.6.13
sqlparse==0.5.3
typing-extensions==4.13.2
whitenoise==6.7.0
